+++ NEWS.md
@@ -10,6 +10,7 @@
 * Better parsing for named indices in `as_draws_rvars()`, including nested use
 of `[`, like `x[y[1],2]` (#243).
 * Allow 0-length `rvar`s with `ndraws() > 1` (#242). 
+* Ensure 0-length `rvar`s can be cast to `draws` formats (#242).
 
 
 # posterior 1.2.2